Creando componentes de Profile
